# Optimize WindowOp's findEndPosition