@feilong-liu 2023-03-10
Why the worst case complexity is O(NlogN)? I still cannot see why it's the case.

Do you have benchmark to show the performance difference?