@HolyLow 2023-03-13
@feilong-liu Thanks for your reply, here is more detailed explanation.

Here is the The O(NlogN) worst case: suppose the original (sorted) data is
```
12345678...N
``` 
And the findEndPosition would be called #N rounds to cut the data into (the data separation points are '^')
```
1^2^3^4^5^6^7^8^...^N^
```
In the 1st round, the input data is `123...N`, and to find the position `1^23...N`, the original algorithm would require $logN$ comparisons as it binary-searches from tail to head.
In the 2nd round, the input data is shorten to `23...N`, and to find the position `2^3...N`, the original algorithm would require $log(N-1)$ comparisons.
...
In the Nth round, the input data is shorten to `N`, and to find the position `N^`, it requires $log1$ comparison.
So the overall comparison is 
$$\sum_{n=1}^{N}logn = O(NlogN)$$

Noted that the comparison operation in WindowOperator is actually quite expensive, because the comparison is based on PagesIndex which is a sorted dataReference for unsorted underlying data.
Here is the benchmark on my local environment (mbp 16 2019, with macOS 12.2.1, intel i7 2.6GHz 6-core CPU, 16GB 2667MHz DDR4).
The benchmark utilizes jmh framework and compares 3 different findEndPosition algorithms: 
- binary search from head to tail (the proposed optimization, indicated by findEndType=0); 
- binary search from tail to head (the original algorithm, indicated by findEndType=1); 
- linear search from head to tail (the basic naive algorithm, indicated by findEndType=2).
The algos operate on a random-generated PagesIndex structure with 0.5 million rows and different partitionSize (indicates the size of each window).
Here is the result:
```
Benchmark                                  (a_arrangeType)  (b_partitionSize)  (c_findEndType)  Mode  Cnt    Score    Error  Units

BenchmarkFindEndPosition.testFindGroupEnd                0                  1                0  avgt    4   14.521 ±  1.494  ms/op
BenchmarkFindEndPosition.testFindGroupEnd                0                  1                1  avgt    4  220.686 ±  8.067  ms/op
BenchmarkFindEndPosition.testFindGroupEnd                0                  1                2  avgt    4   13.256 ±  1.996  ms/op
BenchmarkFindEndPosition.testFindGroupEnd                0                 10                0  avgt    4    9.737 ±  7.999  ms/op
BenchmarkFindEndPosition.testFindGroupEnd                0                 10                1  avgt    4   27.063 ±  9.922  ms/op
BenchmarkFindEndPosition.testFindGroupEnd                0                 10                2  avgt    4   11.901 ±  2.229  ms/op
BenchmarkFindEndPosition.testFindGroupEnd                0                100                0  avgt    4    1.829 ±  0.172  ms/op
BenchmarkFindEndPosition.testFindGroupEnd                0                100                1  avgt    4    3.641 ±  0.207  ms/op
BenchmarkFindEndPosition.testFindGroupEnd                0                100                2  avgt    4   12.134 ±  6.442  ms/op
BenchmarkFindEndPosition.testFindGroupEnd                0               1000                0  avgt    4    0.628 ±  0.064  ms/op
BenchmarkFindEndPosition.testFindGroupEnd                0               1000                1  avgt    4    0.376 ±  0.018  ms/op
BenchmarkFindEndPosition.testFindGroupEnd                0               1000                2  avgt    4   12.305 ±  1.826  ms/op
BenchmarkFindEndPosition.testFindGroupEnd                0             500000                0  avgt    4    0.002 ±  0.001  ms/op
BenchmarkFindEndPosition.testFindGroupEnd                0             500000                1  avgt    4    0.001 ±  0.001  ms/op
BenchmarkFindEndPosition.testFindGroupEnd                0             500000                2  avgt    4   11.482 ±  1.080  ms/op
```
The case with partitionSize=1 is the worst case for the original algorithm (findEndType=1), and the proposed optimization would greatly improve the performance (from 220.7ms/op to 14.5ms/op) under this worst case without hurting the performance in other cases.

Also note that we don't use the naive linear search algorithm (findEndType=2) because it has O(N) complexity in all cases, and would behave worse than the binary searching methods when the data pattern is proper. 
For instance, in the above tests, the naive algorithm becomes the worst algorithm when partitionSize!=1 although it goes well in partitionSize=1.