@wanglinsong 2023-03-27
Please update release notes formatting based on this https://github.com/prestodb/presto/wiki/Release-Notes-Guidelines#formatting.