@aditi-pandit 2023-03-30
> @rohanpednekar @aditi-pandit could you please review this PR or find some proper committer to review it?

I did a quick pass on this code. But I don't have committer rights for this part of the code-base.

@rohanpednekar : Please find the java code owner for window functions also for this PR.