@tdcmeehan 2023-05-08
This change looks good.  I think if you run `BenchmarkWindowOperator` you should find a performance improvement with some of the results.  It would be great to see if we find a performance improvement there, and add the results to this PR and commit message for posterity.